README.md
This program compares an estimate of e as a summation of 1/n! when n>3 and the math value of e.
The function fract_numers is a generator for finding the numerator of the summations.
When plcs is passed to compare_e_est, it searches for the least number of terms to satisfy the requested
decimal place accuracy.
Answering a question such as: What is the minimum number of terms I should use to result in six-place accuracy for e?
Meaning: when will adding more terms not change the first six decimal places?